Allow to run the daemon on slaves
If no replicaSet is configured, and the daemon is executed on a slave,
it ignores reaping. If this slave becomes a master, it starts to reap
again. This allows to make a configuration when only one daemon in the
cluster performs reaping (all other daemons are waiting for their note
to become a new master): to do it, just turn off replicaSet in config
and send UDP packets to a master only from your app.